feature/jdv-streamline-plots (#1341)
* Use Plotly.restyle with a subset of data points each time the user selects zooms into the plot data.
This allows the GUI to be MUCH more responsive for large plot datasets.

* Revert husky pre-commit to 2.y version

* Use lines only for plots zoomed out beyond the point of including all of the data points.

* Refactor JDV source paths
Remove <Router>

* Pull Plots pane out as separate component

* Add some depth to clarify the GUI

* Add selector for the plot type (lines, markers, or auto)
Restyle the plot toolbar

* Update import in JCC

* Add Chart Style label before dropdown

* Continue past the end of the plot's data window by one point, to illustrate that there's more data to the right (if applicable)

* Style the data table better

* Allow clicking on plots as well as hovering, to accomodate tablets better

* Use React state to update plots

* Overflow anywhere for data table values (those mission states can be long)

* typo in bisect.ts comments

"Whish" to "Which"

* Use modulus to eliminate offset jiggling if the user scrubs through the data by dragging the time axis
Include some lower resolution outside data, so the user can see the shape of the adjacent data while scrubbing through

* When all plots are deleted, reset the visible time range

* Add toggle for all data

* Change checkbox to toggle, update name for toggle

---------
