+++ src/web/jdv/client/src/model/Log.ts
