+++ src/web/jdv/client/src/ui/DataTable.css
@@ -1,0 +1,17 @@
+.dataTable {
+    margin: 10pt 0pt 0pt 0pt;
+}
+
+.dataTable > table {
+    width: 100%;
+    padding: 10pt;
+}
+
+.dataTable td {
+    width: 50%;
+    overflow-wrap: anywhere;
+}
+
+table > thead {
+    background-color: aliceblue;
+}
