+++ src/web/jdv/client/src/ui/Plots.css
@@ -1,0 +1,22 @@
+.plotButton {
+    vertical-align: middle;
+    width: auto;
+    padding: 8pt;
+    border: none;
+    border-radius: 4pt;
+    background-color: #0000;
+}
+
+.plotButton:hover {
+    background-color: lightgray;
+}
+
+.plotButtonBar {
+    background-color: #f0f0f0;
+    margin: 4pt;
+
+    label {
+        padding-left: 1em;
+        padding-right: 1em;
+    }
+}
